requirements.txt
# GitHub API и локальный Git
PyGithub>=2.1.1
GitPython>=3.1.0

# LLM: OpenAI (GPT-4o-mini)
openai>=1.12.0

# LLM: YandexGPT (опционально)
# yandex-cloud>=0.300.0
requests>=2.31.0

# Конфигурация
python-dotenv>=1.0.0

# Шифрование секретов для GitHub API (скрипт настройки)
PyNaCl>=1.5.0

# Качество кода (по ТЗ)
ruff>=0.1.0
black>=24.0.0
mypy>=1.8.0
pytest>=8.0.0
